Move User to shared.networking; nest UserInfo; require senderId on Request
- Add shared.networking.User with private nested User.UserInfo and User.userInfo factories
- Remove server.User and payload UserInfo; update imports across server, payloads, client, tests
- Request: only constructors that include senderId; drop no-sender overloads
- NetworkingSeedData: pass senderId on requests; fix loginFailedResponse; drop obsolete directory helpers
- LoginResult: remove duplicate constructor

Made-with: Cursor
